Clean up core objects
This commit applies a number of corrections and changes required
for base functionality. Prior to this commit, nothing was known to
work. Manual smoke testing revealed a number of issues with the
code as it was written.

The manual smoke testing also confirmed that on a fundamental level
the interactions and results are desireable. This code is extremely
sloppy still and totally unproven at any form of scale. Enjoy :)