Add init.pp to manage configuration and gems
This commit adds an initial default class to manage the requirements of
the Adopter.

Later on this class will be extended to manage any needed configurations
in the event this adopter is not installed on the master. For now the
Adopter is relying on a monolithic install and being installed on the
master.