Allow node to lookup by transaction UUID
Prior to this commit, the node object was relying solely on the most
recently posted event data to PuppetDB. While this is likely to be
acceptable, more precise data could be used for lookups. Every Puppet
run generates a UUID that is tied to the report which contains the
events.

This commit allows a precise transaction UUID to be used for lookup of
event data in PuppetDB. If the UUID is not available, the search will
fall back on the most recently submitted data approach.