Allow individual nodes to be pulled from NodeGroup
Without this commit, the node objects discovered by the NodeGroup can
only be accessed in bulk and not individually. The missing behaviour
makes it difficult to specifically work with individual nodes by name if
needed. This makes things such as adding additional information to the
nodes after discovery difficult.

This commit refactors the way nodes are stored inside the NodeGroup
object. The original behaviour are still preserved by not nodes are
stored in a hash. This commit also adds square bracket syntax to the
NodeGroup to access nodes by certname.