Parse PCP message data
When PCP messages are received, the data in them coming back from the
PXP agent is JSON. Prior to this commit, this data was not being parsed
so that it may be properly used. This was causing failures in the Node
object as no usable data was being sent.

This commit parses the JSON data before trying to use it.