Defer lookup of report hash
Without this commit, a Node object will attempt to lookup the report
hash generated from a Puppet Agent run as soon as it receives it,
however, this may not have been enough time for the report to be
ingested by PuppetDB yet. This operation is a little too eager.

This commit holds off on the resolution of the report hash until the
events are being requested. This way PuppetDB has a chance to process
all the necessary data.