Update name of variable in Event class
This commit uses the name "seen_percentage" for the variable that is
used to determine how often an event is seen in catalog. This
seen_percentage is used to determine if a event can be used for a
variation group or if it is too fuzzy to be used in correlation.