Refactor for console group creation, cache refresh
This commit refactors the code to be cleaner and simpler to follow. This
is all towards the end of making it easier to add the ability for the
tool to refresh console and master caches after installing a new module.

Of other interest  the Util class, currently used for keeping reusable
methods that visualize and execute code that blocks. Some duplication
was removed as well.