Create EventSet Class
Currently this doesn't serve a ton of purpose. It's basically identical to the
Set class. What it does give us is the ability to change what an
EventSet is, and what methods you can call on them, without having to do
a ton of other refactoring (like if we switch away from using Set to an
EventSet object later on instead of now). Classes are cheap, so let's
just make one.