Improve settings code
This commit is a majority rewrite of the code that loads and handles
settings in the tool. This commit also handles a method of consistently
accessing code throughout the tool in a manner inspired from the core
Puppet code. This class will be expanded with further settings and
defaults later on.